pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claims-governed-poc"
version = "0.1.0"
description = "Governed Snowflake claims ingestion pipeline PoC"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "snowflake-connector-python>=3.12.0",
  "PyYAML>=6.0.2",
  "jsonschema>=4.23.0",
  "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0"]
dashboard = [
  "streamlit>=1.43.0",
  "pandas>=2.2.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["pipeline*", "scripts*", "config*"]
namespaces = false
